# docs/openapi_backend_service.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/rest/subscriptions":{"get":{"tags":["backend-rest-controller"],"summary":"Gets all subscriptions in the system","operationId":"getAll","responses":{"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"string"}}}},"200":{"description":"Found 1 or more items","content":{"application/json":{"schema":{"type":"string"}}}},"404":{"description":"No items found","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized server error","content":{"application/json":{"schema":{"type":"string"}}}}}},"post":{"tags":["backend-rest-controller"],"summary":"Persists the Subscription in the request body if it was not in the database; if it was already in the system, that one is returned, without any change happenning in the data layer","operationId":"addOne","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscription"}}},"required":true},"responses":{"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"string"}}}},"201":{"description":"Created a new item","content":{"application/json":{"schema":{"type":"string"}}}},"417":{"description":"Found an item that already existed","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized server error","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/rest/subscriptions/{id}":{"get":{"tags":["backend-rest-controller"],"summary":"Gets the subscription with the id specified","operationId":"getEntityById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"string"}}}},"200":{"description":"Found an item","content":{"application/json":{"schema":{"type":"string"}}}},"404":{"description":"No items found","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized server error","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"tags":["backend-rest-controller"],"summary":"Deletes the subscription corresponding to the specified id if present in the system","operationId":"removeEntity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"string"}}}},"404":{"description":"Item not found","content":{"application/json":{"schema":{"type":"string"}}}},"200":{"description":"Deleted the specified item","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized server error","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Subscription":{"required":["birthdate","consent","email","newsletterId"],"type":"object","properties":{"subscriptionId":{"type":"integer","format":"int64"},"email":{"type":"string"},"firstName":{"type":"string"},"gender":{"type":"string"},"consent":{"type":"boolean"},"birthdate":{"type":"string","format":"date"},"newsletterId":{"type":"integer","format":"int64"}}}}}}